Switched the precarious bundle adjustment functions to use SSBA instead of SBA.  Actually added the multiview/bundle* files which i forgot to upload before.
SSBA seems what we need for now.  It does metric BA with constant or varying camera parameters including radial distortion.  It ignores outliers automatically.  It is not "generic" as sba but does what we need and is LGPL.  Kudos go for Christopher Zach.  And thanks to bomboze for the info.