Refactor the robust fitting system.
This changes the way the solvers are presented to the robust fitting system.
It's a big change, and at the moment the homography test is still broken. Going
forward it will make it easier to write new robust estimation algorithms.

Of particular note: See two_view_kernel.h and robust_estimation.h. They
demonstrate the concept of a robust fitting "kernel".

This also introduces namespaces to contain the solving kernels for each
relation.

Note:

* homography_test is still broken; will fix soon.
* There is still dead code around in fundamental / homography; will fix soon.
* The resection, five point, and nview code still needs to get ported to take
  advantage of the new solver system.